parseResponse -> "post".. it doesn't just parse.. it actually dial out to do the request. parseResponseMultipart => postWithMultipartResponse too.
Got rid of the "Config"... we don't need "protocol" and "origin" for the
websocket on Slack, no need to keep that around. So only "token" was left.

Simplified parseResponseBody.. was slightly convoluted.

Attached the "search" to the Slack methods.  This way, we don,t need to pass
the token or the debug flag.  Will do the same with "post()" next.

Moved things around in websocket.go websocket_*.go

Stubbed "manageConnection" connect/disconnect manager.