check the HTTP status code and content type headers
This change checks the response code to make sure the server returned
a 200. When Slack returns a non-200 response code, the body might be
HTML, which causes JSON parsing to fail with:

    invalid character '<' looking for beginning of value

This is misleading since the real problem is the server error which is now
reported in the error.

I saw this locally with file uploads timing out with a 504 Gateway
Timeout that was invisible until I added some debug logs and then this
code.

In a previous patch, I checked the Content-Type as well, but that
quickly proved to be fragile and unnecessary.

Also added a logResponse() function that uses net/http/httputil to dump
the response when in debug mode. net/http/httputil is already used in
websocket_proxy.go so it does not create a new dependency.