Support for reminders.add
This is a straightforward implementation attempting to mirror patterns from
sibling handlers.

The only quirk is the fact that the endpoint actually supports "user" or
"channel" as optional parameters, even though "channel" is not in
Slack's API docs currently. I'm implementing this specifically for use
managing channel reminders so I added it in.