Handle state in interaction callback
The type of state field in a Slack response to interactons now depends
on the type of interaction. Specifically, the correspondence is as
follows.

- type: dialog_submisson -> state: string
- type: block_actions -> state: dictionary array

This commit is trying to handle these properly while maintaining backwards
compatibility, so futureversions may change to a more approproate design.

See Also: https://api.slack.com/changelog/2020-09-full-state-on-view-submisson-and-block-actions

resolved: #816